requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
azure-ai-textanalytics==5.1.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
azure-core==1.20.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
frozenlist==1.2.0
future==0.18.2
idna==3.3
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.3
line-bot-sdk==2.0.1
MarkupSafe==2.0.1
msrest==0.6.21
multidict==5.2.0
oauthlib==3.1.1
Pillow==8.4.0
pygame==2.1.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
